Added some more comments to explain what is happening
