+++ lib/ui/screens/types/modal_draggable.dart
@@ -1,10 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:pokedex/configs/colors.dart';
 import 'package:pokedex/configs/images.dart';
-import 'package:pokedex/ui/screens/types/type_listview.dart';
+import 'package:pokedex/ui/screens/types/modal_contents.dart';
 
-class TypeEffectSheet extends StatelessWidget {
-  const TypeEffectSheet({
+// Class responsible for generating the modal page when clicked on a type
+class ModalDraggable extends StatelessWidget {
+  const ModalDraggable({
     Key key,
     @required this.width,
     @required this.index,
@@ -27,16 +28,6 @@ class TypeEffectSheet extends StatelessWidget {
                     BorderRadius.only(topLeft: Radius.circular(20), topRight: Radius.circular(20))),
             child: Stack(
               children: [
-                // Positioned(
-                //   top: -w / 10,
-                //   left: -w / 6,
-                //   child: Image(
-                //     image: AppImages.pokeball,
-                //     width: w / 2,
-                //     height: w / 2,
-                //     color: AppColors.black.withOpacity(0.1),
-                //   ),
-                // ),
                 Align(
                   child: Image(
                     image: AppImages.pokeball,
@@ -45,17 +36,7 @@ class TypeEffectSheet extends StatelessWidget {
                     color: AppColors.black.withOpacity(0.1),
                   ),
                 ),
-                // Positioned(
-                //   bottom: -w / 10,
-                //   right: -w / 6,
-                //   child: Image(
-                //     image: AppImages.pokeball,
-                //     width: w / 2,
-                //     height: w / 2,
-                //     color: AppColors.black.withOpacity(0.1),
-                //   ),
-                // ),
-                TypeFullList(index: index, width: width, scroller: s),
+                ModalContents(index: index, width: width, scroller: s), //type_listview
               ],
             ),
           );
