+++ lib/ui/screens/types/type_entities/type_constants.dart
