+++ lib/ui/screens/types/type_entities/widget_list.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:pokedex/configs/colors.dart';
-import 'package:pokedex/ui/screens/types/type_card.dart';
+import 'package:pokedex/ui/screens/types/type_container.dart';
 import 'package:pokedex/ui/screens/types/type_entities/type_funcs.dart';
 
-List<Widget> lister(int i, dynamic n, double w, String term) {
-  var arr = effectreturner(i, n);
+// A function that return a list of widgets that are used in displaying the effectiveness
+List<Widget> lister(int index, dynamic value, double width, String term) {
+  var arr = effectreturner(index, value);
   return [
     Container(
       alignment: Alignment.centerLeft,
@@ -20,7 +20,7 @@ List<Widget> lister(int i, dynamic n, double w, String term) {
     Container(
         alignment: Alignment.topCenter,
         padding: const EdgeInsets.only(left: 8.0, top: 8),
-        width: w * 0.75,
+        width: width * 0.75,
         child: Wrap(
           crossAxisAlignment: WrapCrossAlignment.center,
           alignment: WrapAlignment.center,
@@ -29,9 +29,9 @@ List<Widget> lister(int i, dynamic n, double w, String term) {
               Padding(
                 padding: const EdgeInsets.only(top: 8.0),
                 child: TypeDisplayContainer(
-                  i: i,
+                  index: index,
                   path: "effects",
-                  value: n,
+                  value: value,
                   width: 75.0,
                   j: j,
                   height: 25,
@@ -43,7 +43,7 @@ List<Widget> lister(int i, dynamic n, double w, String term) {
     Container(
       margin: EdgeInsets.only(top: 20),
       height: 1,
-      width: w / 1.7,
+      width: width / 1.7,
       color: Colors.black12,
     ),
   ];
