+++ lib/ui/screens/types/type_screen.dart
@@ -1,12 +1,10 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:pokedex/configs/colors.dart';
-import 'package:pokedex/configs/images.dart';
-import 'package:pokedex/configs/types.dart';
 import 'package:pokedex/routes.dart';
 import 'package:pokedex/ui/screens/types/type_grid.dart';
 import 'package:pokedex/ui/widgets/poke_container.dart';
 
+// Class that is related to creating the whole type-effects page
 class TypeEffectScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
@@ -66,7 +64,7 @@ class TypeEffectScreen extends StatelessWidget {
                   ),
                 ),
               ),
-              TypeEffectGrid()
+              TypeEffectGrid() // type_grid
             ],
           ),
         ];
